requirements.txt
aiogram==3.4.1
aiohttp==3.9.3
aiohttp_retry==2.8.3
beautifulsoup4==4.11.1
flake8==7.0.0
mypy==1.8.0
pytest==8.0.0
pytest-asyncio==0.23.5
python-dotenv~=1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0
types_beautifulsoup4==4.12.0